+++ code/flight-computer-lib/src/device/bno055.rs
@@ -103,7 +103,7 @@ where
             gyro,
             mag,
             temperature,
-            timestamp: Time::new::<microsecond>(Instant::now().as_micros()),
+            timestamp: Time::new::<microsecond>(Instant::now().as_micros() as f64),
         })
     }
 }
